:sparkles: add feature do disable and enable all blockers
