+++ src/styles/components/maincontent.style.scss
@@ -31,6 +31,54 @@ main {
             background-color: #5bb06f;
         }
     }
+
+    article.controlsContainer {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        flex-direction: column;
+        width: 100%;
+
+        section.row {
+            padding: 5px 0;
+            display: flex;
+            width: 100%;
+
+            button {
+                width: 100%;
+                padding: 10px 0px;
+                margin: 3px 10px;
+                cursor: pointer;
+                font-weight: bold;
+            }
+
+            button.btnDisableAll {
+                background-color: #722f37;
+                border: solid 2px #fff;
+                border-radius: 5px;
+                transition: 0.3s;
+                color: #fff;
+
+                &:hover {
+                    background-color: #692b32;
+                    transition: 0.3s;
+                }
+            }
+
+            button.btnEnableAll {
+                background-color: #73dc8c;
+                border: solid 2px #fff;
+                border-radius: 5px;
+                transition: 0.3s;
+                color: #000;
+
+                &:hover {
+                    background-color: #5bb06f;
+                    transition: 0.3s;
+                }
+            }
+        }
+    }
 }
 
 details {
@@ -190,67 +238,6 @@ details {
 
                 figcaption {
                     font-weight: bold;
-                }
-            }
-        }
-    }
-
-    article.advancedContainer {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-direction: column;
-
-        section.row {
-            padding: 5px 0;
-
-            input#inputCsvImport {
-                display: none;
-            }
-
-            button {
-                width: 140px;
-                padding: 10px 0px;
-                margin: 10px;
-                cursor: pointer;
-                font-weight: bold;
-            }
-
-            button.btnDisableAll {
-                background-color: #722f37;
-                border: solid 2px #fff;
-                border-radius: 5px;
-                transition: .3s;
-                color: #fff;
-
-                &:hover {
-                    background-color: #692B32;
-                    transition: .3s;
-                }
-            }
-
-            button.btnEnableAll {
-                background-color: #73dc8c;
-                border: solid 2px #000;
-                border-radius: 5px;
-                transition: .3s;
-                color: #000;
-
-                &:hover {
-                    background-color: #5bb06f;
-                    transition: .3s;
-                }
-            }
-
-            button.btnImport {
-                background-color: #f3f3f3;
-                color: #000;
-                border: dotted 2px #000;
-                transition: .3s;
-                border-radius: 5px;
-
-                &:hover {
-                    background-color: #9f9f9f;
                 }
             }
         }
